requirements.txt
anyio==4.9.0
certifi==2025.7.14
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
MouseInfo==0.1.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytchat==0.5.5
python-dotenv==1.1.1
python3-xlib==0.15
pytweening==1.2.0
sniffio==1.3.1
typing_extensions==4.14.1
